Implement Admin functionalities

+++ app/job/job_views.py
@@ -5,7 +5,6 @@ Module for job views
 import os
 import uuid
 from models import db
-from models.user import User
 from werkzeug.utils import secure_filename
 from app.job.job_forms import JobApplicationForm
 from flask import (
@@ -34,7 +33,7 @@ auth = Auth()
 
 
 @job_views.route('/apply-for-job/<string:job_id>', methods=['GET', 'POST'])
-@login_required
+@auth.login_required()
 @auth.role_required('Job Seeker')
 def apply_for_job(job_id: str) -> ResponseReturnValue:
     job = db.find(JobListing, id=job_id)
@@ -58,7 +57,7 @@ def apply_for_job(job_id: str) -> ResponseReturnValue:
                 resume=filename
             )
 
-            db.update(JobListing, job.id, applications_count=job.applications_count+1)
+            db.update(JobListing, job.id)
             flash('Application submitted', 'success')
             return redirect(url_for('job_views.view_job_details', job_id=job_id))
         return render_template('job/job_application.html', form=form, job=job)
@@ -67,7 +66,7 @@ def apply_for_job(job_id: str) -> ResponseReturnValue:
 
 
 @job_views.route('/update-job-application/<string:job_id>', methods=['GET', 'POST'])
-@login_required
+@auth.login_required()
 @auth.role_required('Job Seeker')
 def update_job_application(job_id: str) -> ResponseReturnValue:
     job = db.find(JobListing, id=job_id)
@@ -77,22 +76,30 @@ def update_job_application(job_id: str) -> ResponseReturnValue:
     user_profile = current_user.get_profile()
     application = db.find(JobApplication, job=job, user=user_profile)
     if application:
-        form = JobApplicationForm(cover_letter=application.cover_letter)
+        form = JobApplicationForm()
+        if request.method == 'GET':
+            form = JobApplicationForm(cover_letter=application.cover_letter)
 
         if form.validate_on_submit():
             resume = form.resume.data
-            filename = secure_filename(resume.filename)
-            filename = str(uuid.uuid4()) + filename
-            file_path = os.path.join(current_app.config['RESUME_UPLOAD_FOLDER'], application.resume)
-            if os.path.exists(file_path):
-                os.remove(file_path)
+            if resume:
+                filename = secure_filename(resume.filename)
+                filename = str(uuid.uuid4()) + filename
+                file_path = os.path.join(current_app.config['RESUME_UPLOAD_FOLDER'], application.resume)
+                if os.path.exists(file_path):
+                    os.remove(file_path)
 
-            resume.save(os.path.join(current_app.config['RESUME_UPLOAD_FOLDER'], filename))
+                resume.save(os.path.join(current_app.config['RESUME_UPLOAD_FOLDER'], filename))
+                db.update(
+                    JobApplication,
+                    id=application.id,
+                    resume=filename
+                )
+
             db.update(
                 JobApplication,
                 id=application.id,
                 cover_letter=form.cover_letter.data,
-                resume=filename
             )
 
             flash('Application updated', 'info')
@@ -103,7 +110,7 @@ def update_job_application(job_id: str) -> ResponseReturnValue:
 
 
 @job_views.route('/cancel-job-application/<string:job_id>', methods=['GET'])
-@login_required
+@auth.login_required()
 @auth.role_required('Job Seeker')
 def cancel_job_application(job_id: str) -> ResponseReturnValue:
     job = db.find(JobListing, id=job_id)
@@ -120,7 +127,7 @@ def cancel_job_application(job_id: str) -> ResponseReturnValue:
         application.id
     )
 
-    db.update(JobListing, job.id, applications_count=job.applications_count-1)
+    db.update(JobListing, job.id)
     flash('Application cancelled', 'info')
 
     referer = request.headers.get("Referer")
@@ -129,7 +136,6 @@ def cancel_job_application(job_id: str) -> ResponseReturnValue:
 
 @job_views.route('/view-job/<string:job_id>', methods=['GET'])
 def view_job_details(job_id: str) -> ResponseReturnValue:
-    applied = False
     job = db.find(JobListing, id=job_id)
     if not job:
         abort(404)
@@ -142,23 +148,144 @@ def view_job_details(job_id: str) -> ResponseReturnValue:
             application = db.find(JobApplication, job=job, user=user_profile)
 
         if application:
-            applied = True
+            return render_template('job/job_details.html', job=job, application=application)
 
-    return render_template('job/job_details.html', job=job, applied=applied)
+    return render_template('job/job_details.html', job=job)
 
 
-@job_views.route('/view-applications/<string:job_id>', methods=['GET', 'POST'])
-@login_required
+@job_views.route('/view-applications/<string:job_id>', methods=['GET'])
+@auth.login_required()
 @auth.role_required('Company')
 def view_applications(job_id: str) -> ResponseReturnValue:
     """
     View applications on a job
     """
-    pass
+    company = current_user.get_company()
+    job = db.find(JobListing, id=job_id, company=company)
+    if not job:
+        abort(404)
+
+    applications = db.filter(JobApplication, job=job)
+    return render_template('job/view_job_application.html', job=job, applications=applications)
+
+
+@job_views.route('/view-application-details/<string:application_id>', methods=['GET'])
+@auth.login_required()
+@auth.role_required('Company')
+def view_application_details(application_id: str) -> ResponseReturnValue:
+    company = current_user.get_company()
+    application = db.find(JobApplication, id=application_id)
+    if not application:
+        abort(404)
+
+    if application.job not in db.filter(JobListing, company=company):
+        abort(404)
+
+    if application.status == 'submitted':
+        db.update(JobApplication, application_id, status='under_review')
+
+    application = db.find(JobApplication, id=application_id)
+
+    return render_template('job/view_application_details.html', application=application)
+
+
+@job_views.route('/cancel-interview/<string:application_id>', methods=['GET'])
+@auth.login_required()
+@auth.role_required('Company')
+def cancel_interview(application_id: str) -> ResponseReturnValue:
+    company = current_user.get_company()
+    application = db.find(JobApplication, id=application_id)
+    if not application:
+        abort(404)
+
+    if application.job not in db.filter(JobListing, company=company):
+        abort(404)
+
+    db.update(JobApplication, application_id, status='under_review')
+    flash('Interview canceled', 'info')
+    application = db.find(JobApplication, id=application_id)
+
+    return render_template('job/view_application_details.html', application=application)
+
+
+@job_views.route('/cancel-offer/<string:application_id>', methods=['GET'])
+@auth.login_required()
+@auth.role_required('Company')
+def cancel_offer(application_id: str) -> ResponseReturnValue:
+    company = current_user.get_company()
+    application = db.find(JobApplication, id=application_id)
+    if not application:
+        abort(404)
+
+    if application.job not in db.filter(JobListing, company=company):
+        abort(404)
+
+    db.update(JobApplication, application_id, status='under_review')
+    flash('Offer canceled', 'info')
+    application = db.find(JobApplication, id=application_id)
+
+    return render_template('job/view_application_details.html', application=application)
+
+
+@job_views.route('/schedule-interview/<string:application_id>', methods=['GET'])
+@auth.login_required()
+@auth.role_required('Company')
+def schedule_interview(application_id: str) -> ResponseReturnValue:
+    company = current_user.get_company()
+    application = db.find(JobApplication, id=application_id)
+    if not application:
+        abort(404)
+
+    if application.job not in db.filter(JobListing, company=company):
+        abort(404)
+
+    db.update(JobApplication, application_id, status='interview_scheduled')
+    flash('Interview Scheduled', 'info')
+    application = db.find(JobApplication, id=application_id)
+
+    return render_template('job/view_application_details.html', application=application)
+
+
+@job_views.route('/offer-job/<string:application_id>', methods=['GET'])
+@auth.login_required()
+@auth.role_required('Company')
+def offer_job(application_id: str) -> ResponseReturnValue:
+    company = current_user.get_company()
+    application = db.find(JobApplication, id=application_id)
+    if not application:
+        abort(404)
+
+    if application.job not in db.filter(JobListing, company=company):
+        abort(404)
+
+    db.update(JobApplication, application_id, status='offer_received')
+    flash('Offer sent', 'info')
+    application = db.find(JobApplication, id=application_id)
+
+    return render_template('job/view_application_details.html', application=application)
+
+
+@job_views.route('/reject-application/<string:application_id>', methods=['GET'])
+@auth.login_required()
+@auth.role_required('Company')
+def reject_application(application_id: str) -> ResponseReturnValue:
+    company = current_user.get_company()
+    application = db.find(JobApplication, id=application_id)
+    if not application:
+        abort(404)
+
+    if application.job not in db.filter(JobListing, company=company):
+        abort(404)
+
+    db.update(JobApplication, application_id, status='rejected')
+    flash('Application rejected', 'info')
+    application = db.find(JobApplication, id=application_id)
+
+    return render_template('job/view_application_details.html', application=application)
 
 
 @job_views.route('/save-job/<string:job_id>', methods=['GET'])
-@login_required
+@auth.login_required()
 @auth.role_required('Job Seeker')
 def save_job(job_id: str) -> ResponseReturnValue:
     """
@@ -175,7 +302,7 @@ def save_job(job_id: str) -> ResponseReturnValue:
 
 
 @job_views.route('/unsave-job/<string:job_id>', methods=['GET'])
-@login_required
+@auth.login_required()
 @auth.role_required('Job Seeker')
 def unsave_job(job_id: str) -> ResponseReturnValue:
     """
@@ -192,23 +319,24 @@ def unsave_job(job_id: str) -> ResponseReturnValue:
 
 
 @job_views.route('/view-saved-jobs', methods=['GET'])
-@login_required
+@auth.login_required()
 @auth.role_required('Job Seeker')
 def view_saved_jobs() -> ResponseReturnValue:
     return render_template('job/saved_jobs.html')
 
 
 @job_views.route('/delete-job/<string:job_id>')
-@login_required
-@auth.role_required('Company')
+@auth.login_required()
+@auth.role_required('Admin', 'Company', 'Super Admin')
 def delete_job(job_id: str) -> ResponseReturnValue:
     """
     Delete a job posting
     """
     db.delete(JobListing, id=job_id)
     flash('Job deleted', 'info')
+    referer = request.headers.get('Referer')
 
-    return redirect(url_for('user_views.dashboard'))
+    return redirect(referer or url_for('user_views.dashboard'))
 
 
 @job_views.route('/post-job', methods=['GET', 'POST'])
@@ -219,24 +347,17 @@ def post_job() -> ResponseReturnValue:
     Post a job
     """
     form = PostJobForm()
-    print(form.validate_on_submit())
     if form.validate_on_submit():
-        skills = [skill.data for skill in form.skills]
-
+        skills = form.skills.data.split(', ')
         details = request.form.to_dict()
-        job_details = {
+        _ = details.pop('csrf_token')
+        details.update({
+            'skills': skills,
             'source': 'Company',
             'company': current_user.get_company(),
-            'skills': skills
-        }
+        })
 
-        for key, val in details.items():
-            if key.startswith('skills-') or key == 'csrf_token':
-                continue
-
-            job_details[key] = val
-
-        _ = db.insert(JobListing, **job_details)
+        _ = db.insert(JobListing, **details)
         flash('Job posted', 'success')
 
         return redirect(url_for('user_views.dashboard'))
@@ -245,7 +366,7 @@ def post_job() -> ResponseReturnValue:
 
 
 @job_views.route('/edit-job/<string:job_id>', methods=['GET', 'POST'])
-@login_required
+@auth.login_required()
 @auth.role_required('Company')
 def edit_job(job_id: str) -> ResponseReturnValue:
     """
@@ -256,33 +377,28 @@ def edit_job(job_id: str) -> ResponseReturnValue:
     if not job:
         abort(404)
 
-    form = PostJobForm(
-        title=job.title,
-        description=job.description,
-        skills=job.skills,
-        job_type=job.job_type,
-        location=job.location,
-        salary=job.salary
-    )
+    form = PostJobForm()
+    if request.method == 'GET':
+        skills = ', '.join(job.skills)
+        form = PostJobForm(
+            title=job.title,
+            description=job.description,
+            skills=skills,
+            job_type=job.job_type,
+            location=job.location,
+            salary=job.salary
+        )
 
     if form.validate_on_submit():
-        skills = [skill.data for skill in form.skills]
+        skills = form.skills.data.split(', ')
 
         details = request.form.to_dict()
-        job_details = {
-            'skills': skills
-        }
-
-        for key, val in details.items():
-            if key.startswith('skills-') or key == 'csrf_token':
-                continue
-
-            job_details[key] = val
-
+        _ = details.pop('csrf_token')
+        details['skills'] = skills
         db.update(
             JobListing,
             id=job.id,
-            **job_details
+            **details
         )
 
         flash('Job updated', 'info')
